Release v1.5.6-beta4: security & static-analysis hardening
Static analysis to zero across the backend (staticcheck, golangci-lint,
gosec); go vet (nilness), full go test ./... incl. -race, govulncheck, and
the frontend lint/typecheck/unit suites all green. Migrate deprecated
sing/common/atomic.Int64 -> sync/atomic and net.Error.Temporary() -> a
timeout-based check; remove dead code; make previously unchecked error
returns explicit.

Security/contract hardening (five deferred XFAIL tests now enforced):
- APIv2 invalid/expired bearer token returns HTTP 401 instead of 200
  success=false; browser UI unaffected (cookie sessions on /api, not
  bearer tokens on /apiv2). External API consumers must check HTTP status.
- New opt-in sessionSameSiteStrict setting (default off) -> SameSite=Strict
  session cookies.
- SSRF: reject userinfo in Telegram proxy URLs; reject private/loopback/
  link-local/multicast IP literals in optional HTTP URL settings.
- Harden constant-time API-token-scope comparison against truncation when
  the length difference is a multiple of 256.
- Emit settings_save_succeeded audit event on successful settings saves.
- Enforce backup -> restore table-count consistency (incl. no-TLS sentinel).

Bump config/version and frontend package(-lock).json to 1.5.6-beta4; add
docs/releases/v1.5.6-beta4.md; update EN/RU/ZH changelogs and README.
Untrack the internal docs/project-context.md dev log (not needed by users).

+++ service/server.go
@@ -161,7 +161,7 @@ func (s *ServerService) GetSingboxInfo() map[string]interface{} {
 	return map[string]interface{}{
 		"running": isRunning,
 		"stats": map[string]interface{}{
-			"NumGoroutine": uint32(runtime.NumGoroutine()),
+			"NumGoroutine": runtime.NumGoroutine(),
 			"Alloc":        rtm.Alloc,
 			"Uptime":       uptime,
 		},
@@ -174,7 +174,7 @@ func (s *ServerService) GetSystemInfo() map[string]interface{} {
 	runtime.ReadMemStats(&rtm)
 
 	info["appMem"] = rtm.Sys
-	info["appThreads"] = uint32(runtime.NumGoroutine())
+	info["appThreads"] = runtime.NumGoroutine()
 	cpuInfo, err := cpu.Info()
 	if err == nil && len(cpuInfo) > 0 {
 		info["cpuType"] = cpuInfo[0].ModelName
